Clean up components, separate into molecular, styles
