+++ front-end/src/components/Pages/Billing.js
@@ -1,10 +1,6 @@
 import React, { Component } from "react";
-import { connect } from "react-redux";
-import { Redirect } from "react-router-dom";
-import Sidebar from "../Organisms/SideNav.js";
-
-
-import { Container, FormItem, Form } from "../../styles/signin.js";
+// TODO: MAKE OWN STYLES
+import { BillingContainer, FormItem, Form } from "../../styles/Signin.js";
 
 class Billing extends Component {
   state = {
@@ -15,7 +11,7 @@ class Billing extends Component {
 
   submitHandler = event => {
     event.preventDefault();
-    console.log(event.target.value);
+    // TODO: HANDLE STRIPE
   };
 
   inputChangeHandler = event => {
@@ -25,57 +21,41 @@ class Billing extends Component {
 
   render() {
     return (
-      <div>
-        <Container>
-        <div classname="main-container">
-          <div>
-            <Form onSubmit={this.submitHandler}>
-              <div className="card_body">
-              <FormItem>
-                <div>
-                  <h3>Credit Card</h3>
-                  <input
-                    value={this.state.CC}
-                    onChange={this.inputChangeHandler}
-                    name="CC"
-                    type="text"
-                  />
-                </div>
+      <BillingContainer>
+        {/* Pull out into Molecular components */}
+        <Form onSubmit={this.submitHandler}>
+          <FormItem>
+            <h3>Credit Card</h3>
+            <input
+              value={this.state.CC}
+              onChange={this.inputChangeHandler}
+              name="CC"
+              type="text"
+            />
+          </FormItem>
+          <FormItem>
+            <h3>Expires on</h3>
+            <input
+              value={this.state.EXP}
+              onChange={this.inputChangeHandler}
+              name="EXP"
+              type="text"
+            />
+          </FormItem>
 
-                </FormItem>
-              </div>
-              <div className="card_body">
-                <FormItem>
-                    <div>
-                      <h3>Expires on</h3>
-                      <input
-                        value={this.state.EXP}
-                        onChange={this.inputChangeHandler}
-                        name="EXP"
-                        type="text"
-                      />
-                    </div>
-                  </FormItem>
-              </div>
-              <div className="card_body">
-                <FormItem>
-                    <div>
-                      <h3>CCV</h3>
-                      <input
-                        value={this.state.CCV}
-                        onChange={this.inputChangeHandler}
-                        name="CCV"
-                        type="text"
-                      />
-                    </div>
-                  </FormItem>
-              </div>
-              <button className="billing_btn">Buy Now</button>
-            </Form>
-          </div>
-        </div>
-        </Container>
-      </div>
+          <FormItem>
+            <h3>CCV</h3>
+            <input
+              value={this.state.CCV}
+              onChange={this.inputChangeHandler}
+              name="CCV"
+              type="text"
+            />
+          </FormItem>
+
+          <button className="billing_btn">Buy Now</button>
+        </Form>
+      </BillingContainer>
     );
   }
 }
