+++ front-end/src/components/Pages/Landing.js
@@ -4,7 +4,8 @@ import "../../styles/LandingRef.css";
 import Modal from "../Organisms/modal";
 
 const Landing = () => {
-  //TODO: Restyle all this
+  // TODO: Restyle all this
+  // TODO: STYLED COMPONENTS
   return (
     <div className="container">
       <div className="background-holder">
