+++ front-end/src/styles/Employee.js
@@ -1,38 +1,37 @@
 import styled from "styled-components";
+// Maybe combine with all other "cards" styling
 
-export const Top = styled.div`
-
-`;
+export const Top = styled.div``;
 export const NavBar = styled.div`
-    display: flex;
-    padding: 1em;
+  display: flex;
+  padding: 1em;
 `;
 
 export const Bottom = styled.div`
-    display: flex;
+  display: flex;
 `;
 
 export const Left = styled.div`
-    padding-left: .5em;
-    padding-right: .5em;
+  padding-left: 0.5em;
+  padding-right: 0.5em;
 `;
 
 export const Right = styled.div`
-    display: flex;
-    padding-left: 1em;
+  display: flex;
+  padding-left: 1em;
 `;
-export const EmployeeCard = styled.div`
-    border: 1px solid black;
-    font-size: .5em;
+export const EmployeeCardContainer = styled.div`
+  border: 1px solid black;
+  font-size: 0.5em;
 `;
 
 export const CardInner = styled.div`
-    border: 1px solid black;
-    max-width: 80%;
-    margin-left: 16px;
-    margin-bottom: 16px;
+  border: 1px solid black;
+  max-width: 80%;
+  margin-left: 16px;
+  margin-bottom: 16px;
 `;
 
 export const NewEmployeeCard = styled.div`
-    border: 1px solid black;
+  border: 1px solid black;
 `;
