+++ front-end/src/styles/Form.js
@@ -1,3 +1,3 @@
 import styled from "styled-components";
 
-// STYLES FOR CALENDAR COMPONENT
+// STYLESHEET FOR ALL FORMS
