+++ front-end/src/styles/LandingRef.css
@@ -1,3 +1,4 @@
+/* Turn into styled component */
 .background-holder {
   background-color: grey;
   width: 100%;
