--- front-end/src/styles/WelcomeRef.css
@@ -1,34 +1,0 @@
-.navbar {
-    display: flex;
-    padding: 1em;
-}
-
-.bottom {
-    display: flex;
-}
-
-.left {
-    padding-left: .5em;
-    padding-right: .5em;
-}
-
-.right {
-    display: flex;
-    padding-left: 1em;
-}
-
-.employee-card {
-    border: 1px solid black;
-    font-size: .5em;
-}
-
-.card-inner {
-    border: 1px solid black;
-    max-width: 80%;
-    margin-left: 16px;
-    margin-bottom: 16px;
-}
-
-.new-employee-card {
-    border: 1px solid black;
-}
