+++ front-end/src/styles/billing.css
@@ -1,3 +1,4 @@
+/* DEPRECIATED -- TURN INTO STYLED COMPONENT */
 .main-container {
   opacity: 0.8;
   position: fixed;
@@ -25,7 +26,3 @@
   margin-left: 30px;
   border-radius: 35px;
 }
-
-/* .billing_btn {
-  
-} */
