--- front-end/src/styles/settings.css
@@ -1,18 +1,0 @@
-.card {
-    display: flex;
-    width: 35%;
-    margin: auto;
-    border: .5px solid black;
-    justify-content: center;
-    border-radius: 5%;
-}
-
-.card_body {
-    justify-content: center;
-    opacity: 1;
-    width: 100%;
-    margin-top: 25px;
-    margin-right: 30px;
-    margin-bottom: 25px;
-    margin-left: 30px;
-}
