--- front-end/src/styles/sidebar.css
@@ -1,9 +1,0 @@
-.sidebar-container {
-  display: flex;
-  border: solid black 1px;
-  flex-direction: column;
-  justify-content: flex-start;
-  width: 20%;
-  height: 80%;
-  margin: 0 5%;
-}
